Refactor output paths and organize modules
